+++ lib/x509name.h
@@ -46,7 +46,7 @@ class x509name
 		QString getMostPopular() const;
 		QString taggedValues() const;
 		QString hash() const;
-		unsigned long hashNum() const;
+		unsigned hashNum() const;
 		bool search(const QRegExp &pattern) const;
 };
